feat: wire up slippage popover to limit order quote
